requirements.txt
Django==4.1.1
djangorestframework==3.14.0
numpy==1.23.3
requests==2.28.1
